api/requirements.txt
# Hotel Metasearch API — Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
boto3>=1.34.0
python-dotenv>=1.0.0
httpx>=0.25.0
curl_cffi>=0.7.0
pydantic>=2.0.0

